Improve R interface to the C++ sampling function
I created a new R interface to the C++ sampling function that gives the
user more flexibility and control. They no longer have to pre-process
the responses, but do have to provide the values corresponding to yea
votes and nay votes if they are other than 0 and 1 or -1 and 1
respectively. The user can now also specify the scale factor and length
scale for the GP's covariance function.

I also had to fix a small typo on the Authors@R field in DESCRIPTION