Allow missing responses
I added a check for missing values in the log likelihood function,
and allowed for the user to specify which values should be treated
as missing in the R interface to the sampler. I also allowed for
multiple values to be specified as corresponding to a 'yea', 'nay',
or missing response.