Setup Travis CI
The first step in ensuring Mac support is to set up Travis CI for OSX.
As part of that, I also added in an example for gpirtMCMC(), though we
may need to change up that example as the API evolves (not having any
examples causes a WARNING during R CMD check).

Later I'll start on the configure script to get things working on Mac.