Add configure
I've set up our first effort at a configure script to ensure OpenMP
is properly handled across systems. It's my own twist on the ideas
behind RcppArmadillo's configure -- I just basically reworked it from
scratch to make sure I knew exactly what was going on, and also so it
was written in a way that was more explicit to me. We'll see if this
flies on Travis, and if so whether it seems to work generally across
various machines available to us.