Move to drawing fstar individually
On my first stab at the sampler, I just drew the predicted values of
the f_j for all possible theta values at once. However, of course it
is much faster to draw a predicted value for f_j at *each* of the
possible values of theta individually, per our Slack discussion. This
update does that, and on a toy dataset with 100 respondents and 100
items, led to a 6x speedup.