Add data processing functions
We now have a class 'response_matrix' that should be fairly robust in
terms of making sure that the data passed by the user is put into a
numeric matrix with only 1, -1, and NA values. There are still likely
several cases that aren't handled yet, but this should handle matrices
and dataframes just fine (see the examples).