Move to probit likelihood
I went ahead and move things to a probit likelihood, and implemented
the algorithm changes that allows us to make to speed things up. It
did run about twice as fast, although it didn't seem to perform quite
as well. Part of this may be due to the fact that I haven't added in
the informative priors, while I did take off the half normal prior on
the first theta. We'll see what happens when we add in the informative
priors.