Require party afilliation specification
Initially, we were restricting the sign of one respondent's ideology
parameter to identify the directionality of the space. This sort of
strategy is nice in a setting with symmetry because you can either
impose this restriction during processing, or in a post-processing
step. However, since we aren't assuming much about the mapping from
ideology to responses, we can't do that. It makes much more sense to
simply allow the user to specify which respondents should be on the
left and which should be on the right. This is not too restrictive a
strategy -- we just change the mean of the prior on the respondent's
ideology parameter accordingly (for right now, either -1 or 1). So,
someone who we think a priori should be on the right *could* end up
on the left.

I also allow here for the user to specify initial values for the
ideology parameters. If not supplied, they are drawn from the prior
distributions.