Allow prior mean specification by party
I added a lot of functionality here to allow specifying prior means and
initial theta values. Ostensibly the approach in the response_matrix
creation function should be flexible enough to expand easily to allow
for more than binary response if/when we want that to happen later down
the road.