Provide data in package and remove GPML IRF plotting
Rather than download data every time we build the vignette, we now
provide a static version of the Senate rollcall data that has all the
votes from the first session of the 116th Congress. The code to create
the data in the package is in data-raw/

I removed the IRF plotting capabilities that relied on gpmlr. I'll
add IRF plotting capability back in later, but it will be based on the
mean of the f* draws.