Implement online IRF estimation and restrict θ prior
Rather than have the user choose whether or not to store the entire set
of f* draws, which honestly can take up quite a bit of RAM, we now
simply calculate the mean of those draws online and return the IRF
estimates no matter what (which is both memory & computationally
efficient).

I also removed the feature which previously allowed users to set prior
means by group. It just seems to be more complicated than its worth; it
should be really easy to add back in if it becomes a commonly requested
feature.